+++ app/crud/crud_investment.py
@@ -73,14 +73,16 @@ class CRUDInvestment:
         db.refresh(db_obj)
         return db_obj
     
-    def delete(self, db: Session, *, investment_id: uuid.UUID) -> Investment:
+    def delete(self, db: Session, *, investment_id: uuid.UUID) -> bool:
         """Delete investment."""
         obj = db.query(Investment).get(investment_id)
-        db.delete(obj)
-        db.commit()
-        return obj
+        if obj:
+            db.delete(obj)
+            db.commit()
+            return True
+        return False
     
-    def exists(self, db: Session, investment_id: uuid.UUID) -> bool:
+    def is_exists(self, db: Session, investment_id: uuid.UUID) -> bool:
         """Check if investment exists."""
         return db.query(Investment).filter(Investment.id == investment_id).first() is not None
